2026-02-17-automating-aitw/pyproject.toml
[project]
name = "automating-aitw"
version = "0.1.0"
description = "Automation tools for AI That Works podcast"
requires-python = ">=3.11"
dependencies = [
    "baml-py==0.220.0",
    "python-dotenv>=0.9.9",
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    "requests>=2.32.5",
    "playwright>=1.57.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"



